add draft-lxin-quic-socket-apis.xml
This document describes a mapping of In-kernel QUIC Implementations into a
sockets API. The benefits of this mapping include compatibility for TCP
applications, access to new QUIC features, and a consolidated error and
event notification scheme. In-kernel QUIC enables usage for both userspace
applications and kernel consumers.

The online tool https://author-tools.ietf.org/ can be used to convert this
xml files to html/pdf/txt files for easier reading.

Signed-off-by: Xin Long <lucien.xin@gmail.com>